microservice/requirements.txt
fastapi
uvicorn
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyjwt